🐛 Fix extension icon switch
