+++ src/lib/services/authorization.ts
@@ -3,6 +3,7 @@ import { getDuration } from '../utility/utils';
 import { userProfile } from '$lib/stores/data';
 import { ProfileUtil } from '$lib/utility';
 import { profileController } from '$lib/controllers/profile.controller';
+import { browserController } from '$lib/controllers';
 
 const accept = async (accept: boolean, domain: string, choice: number = 0) => {
 	try {
@@ -31,6 +32,8 @@ const accept = async (accept: boolean, domain: string, choice: number = 0) => {
 				return x;
 			}
 		});
+
+		browserController.getCurrentTab().then((tab) => browserController.switchIcon({ tabId: tab.id as number }));
 
 		profileController.saveProfile(get(userProfile));
